% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_article_pkg.R
\name{add_article}
\alias{add_article}
\title{Add article to existing project
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
add_article(
  name,
  path = getwd(),
  template_pkg = "Rmdx",
  template_fn = "rmdx_paper"
)
}
\arguments{
\item{name}{name of the article}

\item{path}{path to the project}

\item{template_pkg}{template package to use}

\item{template_fn}{template function to use}
}
\value{
\link{NULL}
}
\description{
Add article to existing project
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
