% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_article_pkg.R
\name{add_article_}
\alias{add_article_}
\title{Add article to existing project
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
add_article_(path, pkg, fn, name, use_home)
}
\arguments{
\item{path}{path to the project}

\item{pkg}{template package to use}

\item{fn}{template function to use}

\item{name}{name of the article}

\item{use_home}{indicates if home page is used to display the article online version}
}
\value{
\link{NULL}
}
\description{
Add article to existing project
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
