% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_article_pkg.R
\name{create_article_pkg}
\alias{create_article_pkg}
\title{Create an article package structure}
\usage{
create_article_pkg(path, template = Rmdx::rmdx_paper, use_home = TRUE, ...)
}
\arguments{
\item{path}{where project will be located}

\item{template}{base document to use. You can go with the default. Or set any
of the \href{https://pkgs.rstudio.com/rticles/}{rticles} formats.
Theoretically any format that follows the conventions of rticles can be used.}

\item{use_home}{indicates if home page is used to display the article online version}

\item{...}{other arguments passed to \code{usethis::\link[usethis:create_package]{create_package}}}
}
\value{
\link{NULL}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
create_article_pkg("~/foofactors")
}
}
