% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_paper_pkg.R
\name{create_paper_pkg}
\alias{create_paper_pkg}
\title{Create an article package structure}
\usage{
create_paper_pkg(
  path,
  template = Rmdx::rmdx_paper,
  use_renv = TRUE,
  use_renv_args = list(),
  ...
)
}
\arguments{
\item{path}{where project will be located}

\item{template}{base document to use. You can go with the default. Or set any
of the \href{https://pkgs.rstudio.com/rticles/}{rticles} formats.
Theoretically any format that follows the conventions of rticles can be used.}

\item{use_renv}{controls whether \href{https://rstudio.github.io/renv/index.html}{renv} is used in development}

\item{use_renv_args}{aditional arguments passed to \code{renv::\link[renv:init]{init}}}

\item{...}{other arguments passed to \code{usethis::\link[usethis:create_package]{create_package}}}
}
\value{
\link{NULL}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
  create_paper_pkg("~/foofactors")
}
}
