% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{knit}
\alias{knit}
\title{Custom knit function for articlepkg}
\usage{
knit(input, ...)
}
\arguments{
\item{input}{The input file to be rendered}

\item{...}{other arguments passed to \code{rmarkdown::\link[rmarkdown:render]{render}}}
}
\value{
the compiled document
}
\description{
Custom knit function for articlepkg
}
\examples{
\dontrun{
---
...
knit: articlepkg::knit
---
}
}
