% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{knit}
\alias{knit}
\title{Custom knit function for articlepkg
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
knit(input, ...)
}
\arguments{
\item{input}{The input file to be rendered}

\item{...}{other arguments passed to \code{rmarkdown::\link[rmarkdown:render]{render}}}
}
\value{
the compiled document
}
\description{
Custom knit function for articlepkg
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
---
...
knit: articlepkg::knit
---
}
}
